requirements.txt
requests
pyyaml
twikit
